+++ pkg/xds/envoy/routes/routes_suite_test.go
@@ -1,0 +1,13 @@
+package routes_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+)
+
+func TestRoutes(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Routes Suite")
+}